pyproject.toml
[tool.poetry]
package-mode = false
name = "Echo"
version = "4.1.0"
description = "A collection of discord bots for moderation, reference sheet management, and scam prevention."
authors = ["Autumn <EchoBot@Autumna1Equin0x.pet>"]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
watchdog = "^6.0.0"
filelock = "^3.17.0"
discord-py = "^2.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
